Extract ImageLabellerWindow from ImageLabeller
